ADD WORKING SOLUTION FOR EX00 ASM:
Adding a working solution for asm on ex00.
Though something is not as good as it could be.
Using the exit wrapper from libc to fix an issue
Apparently with stdio buffers not being flushed

The consequences are weird, when using the exit syscall the output
cannot be redirected in a file.

Will investigate!