WIP OPEN AND WRITE IN ASM:
This is a sandbox commit in grace.s
I went back to the roots to make sure I was able to open a file and
write in it in asm as this is the first time I have to do that.

Working my way up to complete the task. Writing readable code before
obfuscation