ADD WORKING SOLUTION FOR GRACE.S:
Fixing the huge stack mistake I made resulting in a working solution.
This problem got me puzzled for a while, because of a huge mistake
I made and the confusion resulting with the stack misunderstanding...
I usually squash WIP commits, to hide all this, but the process often
results in days spent without commits shown on GH...

Here I am being transparent on my mistake, that's how you learn!