WIP SULLY (ASM):
Work in progress for Sully in assembler x86_64
For now the fork seems to work, so does the open.
Trying to work on execve. Seems to be working too.

The compilation is kept to a C compilation and needs to be changed
to a nasm compilation
Not yet made into a quine.