WORKING SOLUTION FOR SULLY IN ASM:
Finally.
After dreaming the solution, it was easier to fix everything else.

The main issue was a misconception (don't know where it came from)
with execve. I somehow built everything just as if execve was a shell

So indeed it could not work. The easiest work around with the existing
code was to change the first parameter to make it actually use a shell.

This is dirty but it works to comply with the subject and I don't want
to waste too much time on this.