ADD WORKING SOLUTION SULLY (PYTHON):
Adding a working solution for Sully in python.

This puts this project to an end.
I could have indeed modified a few things here or there based on
the knowledge acquired during the whole process. But as all good
things need to end, I shall focus on some other projects!